Move restart_jobs from base service.
Method is only used by pipeline, obs and credentials services.
