+++ mash/services/mash_service.py
@@ -16,7 +16,6 @@
 # along with mash.  If not, see <http://www.gnu.org/licenses/>
 #
 
-import json
 import logging
 import os
 import smtplib
@@ -200,19 +199,6 @@ class MashService(object):
         self.channel.basic.consume(
             callback=callback, queue=queue
         )
-
-    def restart_jobs(self, callback):
-        """
-        Restart jobs from config files.
-
-        Recover from service failure with existing jobs.
-        """
-        for job_file in os.listdir(self.job_directory):
-            with open(os.path.join(self.job_directory, job_file), 'r') \
-                    as conf_file:
-                job_config = json.load(conf_file)
-
-            callback(job_config)
 
     def set_logfile(self, logfile):
         """
