+++ test/unit/services/base/service_test.py
@@ -1,5 +1,3 @@
-import io
-
 from unittest.mock import patch
 from unittest.mock import call
 from unittest.mock import MagicMock, Mock
@@ -11,8 +9,6 @@ from mash.mash_exceptions import (
     MashRabbitConnectionException,
     MashLogSetupException
 )
-
-open_name = "builtins.open"
 
 
 class TestBaseService(object):
@@ -97,23 +93,6 @@ class TestBaseService(object):
         self.service.close_connection()
         self.connection.close.assert_called_once_with()
         self.channel.close.assert_called_once_with()
-
-    @patch('mash.services.mash_service.json.load')
-    @patch('mash.services.mash_service.os.listdir')
-    def test_restart_jobs(self, mock_os_listdir, mock_json_load):
-        self.service.job_directory = 'tmp-dir'
-        mock_os_listdir.return_value = ['job-123.json']
-        mock_json_load.return_value = {'id': '1'}
-
-        with patch(open_name, create=True) as mock_open:
-            mock_open.return_value = MagicMock(spec=io.IOBase)
-            mock_callback = Mock()
-            self.service.restart_jobs(mock_callback)
-
-            file_handle = mock_open.return_value.__enter__.return_value
-            file_handle.read.call_count == 1
-
-        mock_callback.assert_called_once_with({'id': '1'})
 
     def test_unbind_queue(self):
         self.service.unbind_queue(
