+++ test/unit/services/obs/service_test.py
@@ -312,3 +312,20 @@ class TestOBSImageBuildResultService(object):
                     "job_file": "tmp-dir/job-1.json"
                 })
             )
+
+    @patch('mash.services.obs.service.json.load')
+    @patch('mash.services.obs.service.os.listdir')
+    def test_restart_jobs(self, mock_os_listdir, mock_json_load):
+        self.obs_result.job_directory = 'tmp-dir'
+        mock_os_listdir.return_value = ['job-123.json']
+        mock_json_load.return_value = {'id': '1'}
+
+        with patch('builtins.open', create=True) as mock_open:
+            mock_open.return_value = MagicMock(spec=io.IOBase)
+            mock_callback = Mock()
+            self.obs_result.restart_jobs(mock_callback)
+
+            file_handle = mock_open.return_value.__enter__.return_value
+            file_handle.read.call_count == 1
+
+        mock_callback.assert_called_once_with({'id': '1'})
