+++ test/unit/services/pipeline/service_test.py
@@ -845,6 +845,23 @@ class TestPipelineService(object):
                 })
             )
 
+    @patch('mash.services.pipeline_service.json.load')
+    @patch('mash.services.pipeline_service.os.listdir')
+    def test_restart_jobs(self, mock_os_listdir, mock_json_load):
+        self.service.job_directory = 'tmp-dir'
+        mock_os_listdir.return_value = ['job-123.json']
+        mock_json_load.return_value = {'id': '1'}
+
+        with patch('builtins.open', create=True) as mock_open:
+            mock_open.return_value = MagicMock(spec=io.IOBase)
+            mock_callback = Mock()
+            self.service.restart_jobs(mock_callback)
+
+            file_handle = mock_open.return_value.__enter__.return_value
+            file_handle.read.call_count == 1
+
+        mock_callback.assert_called_once_with({'id': '1'})
+
     def test_service_start_job(self):
         job = Mock()
         self.service.jobs['1'] = job
